Refactor Tests
Refactored the E2E tests with Page Object design pattern, and retaining only the core test logic inside the test spec and moving all the page related actions along with the locators to the page objects.

Also implemented is the test data which would help in passing the test data from a JSON file rather than having it hardcoded within the test. This would eliminate the dependency on test data as well as the validation messages